requirements.txt
# FastAPI và API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
apscheduler==3.10.4
python-dotenv==1.0.0

# Scrapy và Playwright dependencies
scrapy>=2.11.0
scrapy-playwright>=0.0.33